+++ AffiseModule/AdService/Classes/usecase/AdServiceUseCase.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+protocol AdServiceUseCase {
+    
+    func initialize()
+
+    func getAttributionToken() -> String?
+}