+++ AffiseModule/AdService/Classes/usecase/AdServiceUseCaseImpl.swift
@@ -1,0 +1,18 @@
+import Foundation
+import AdServices
+
+
+class AdServiceUseCaseImpl : AdServiceUseCase {
+
+    var attributionToken: String?
+
+    func initialize() {
+        if #available(iOS 14.3, *) {
+            attributionToken = try? AAAttribution.attributionToken()
+        }
+    }
+
+    func getAttributionToken() -> String? {
+        return attributionToken
+    }
+}